DGKernels works with the tagging system
Refs #9669

+++ framework/include/base/Assembly.h
@@ -592,12 +592,12 @@ public:
 
   DenseVector<Number> & residualBlock(unsigned int var_num, TagID tag_id = 0)
   {
-    return _sub_Re[static_cast<unsigned int>(tag_id)][var_num];
+    return _sub_Re[tag_id][var_num];
   }
 
   DenseVector<Number> & residualBlockNeighbor(unsigned int var_num, TagID tag_id = 0)
   {
-    return _sub_Rn[static_cast<unsigned int>(tag_id)][var_num];
+    return _sub_Rn[tag_id][var_num];
   }
 
   DenseMatrix<Number> & jacobianBlock(unsigned int ivar, unsigned int jvar, TagID tag = 0);
